+++ src/v/cluster/controller.cc
@@ -18,10 +18,10 @@
 #include "cluster/partition_leaders_table.h"
 #include "cluster/partition_manager.h"
 #include "cluster/raft0_utils.h"
+#include "cluster/reflection.h"
 #include "cluster/shard_table.h"
 #include "cluster/topic_table.h"
 #include "cluster/topics_frontend.h"
-#include "cluster/types.h"
 #include "model/metadata.h"
 #include "model/timeout_clock.h"
