+++ src/v/cluster/controller_backend.cc
@@ -13,9 +13,9 @@
 #include "cluster/logger.h"
 #include "cluster/members_table.h"
 #include "cluster/partition_leaders_table.h"
+#include "cluster/reflection.h"
 #include "cluster/shard_table.h"
 #include "cluster/topic_table.h"
-#include "cluster/types.h"
 #include "config/configuration.h"
 #include "model/fundamental.h"
 #include "model/metadata.h"
