+++ src/v/cluster/tests/configuration_change_test.cc
@@ -7,10 +7,10 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
+#include "cluster/reflection.h"
 #include "cluster/simple_batch_builder.h"
 #include "cluster/tests/cluster_test_fixture.h"
 #include "cluster/tests/utils.h"
-#include "cluster/types.h"
 #include "model/metadata.h"
 #include "test_utils/fixture.h"
 #include "utils/unresolved_address.h"
