+++ src/v/cluster/tests/metadata_dissemination_test.cc
@@ -8,9 +8,9 @@
 // by the Apache License, Version 2.0
 
 #include "cluster/metadata_cache.h"
+#include "cluster/reflection.h"
 #include "cluster/simple_batch_builder.h"
 #include "cluster/tests/cluster_test_fixture.h"
-#include "cluster/types.h"
 #include "model/fundamental.h"
 #include "model/metadata.h"
 #include "model/timeout_clock.h"
