+++ src/v/cluster/tests/topic_updates_dispatcher_test.cc
@@ -8,6 +8,7 @@
 // by the Apache License, Version 2.0
 
 #include "cluster/partition_allocator.h"
+#include "cluster/reflection.h"
 #include "cluster/tests/topic_table_fixture.h"
 #include "cluster/topic_updates_dispatcher.h"
 #include "model/metadata.h"
