+++ src/v/cluster/types.cc
@@ -7,8 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
-#include "cluster/types.h"
-
+#include "cluster/reflection.h"
 #include "model/fundamental.h"
 #include "model/metadata.h"
 #include "tristate.h"
